Percent is an int, not a float

+++ src/Nest/Cluster/NodesStats/NodeStats.cs
@@ -88,7 +88,7 @@ namespace Nest
 		[JsonProperty("load_average")]
 		public float LoadAverage { get; internal set; }
 		[JsonProperty("percent")]
-		public float Percent { get; internal set; }
+		public int Percent { get; internal set; }
 
 		[JsonProperty("mem")]
 		public ExtendedMemoryStats Memory { get; internal set; }
